+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 - Fixed assign_callback error in confluentkafka input
 - Fixed error logging in ` _get_configuration`, which caused the github checks to fail
 - Resolved `mypy` errors in `BaseProcessorTestCase.` by ensuring `self.object` and `self.patchers` are not `None` before accessing attributes.
+- Fix domain resolver errors for invalid domains
+
 
 ## 16.1.0
 ### Deprecations